Added exodm_db_session to authenticate and keep track of active users
exodm_db_user now hashes passwords using bcrypt. This is a very safe
(but also slow) way to ensure that passwords cannot be cracked by
stealing the password file.

The new process, exodm_db_session, uses bcrypt to verify the password
the first time a session is authenticated. It then caches the session
info, together with a SHA hash based on the password and the bcrypt
hash. Subsequent attempts to authenticate are checked by hashing the
offered password with the bcrypt hash, and comparing with the cached hash.
This is fast, and the simple hash is not stored persistently.

Since the safe check takes some time (ca 6-700 ms), any authentication
requests against the same user id that come in during this time are
buffered. If they offered the same password as the request just verified
safely, they are quickly granted (or rejected) once the verification is
done. If the first request failed, and pending requests used a different
password, the safe verification is repeated on the different password,
and so on. This process should be load-regulated.