Registration page in web UI, counter for new UIDs, exodm_db session mgr
The registration page does simple verification of user info, then
stores the data with exodm_db_user. We should do standard validation of
the email address (send a verification email). This would require us to
flag a user account as yet unverified - we don't do this today.

The exodm_db_session session manager validates username and password
against the bcrypt-generated hash in the user database. This is slow,
so once verified, the session manager makes a SHA hash for subsequent
verification. This hash exists in memory until the user logs out, and
mustn't be stored persistently (this would void the value of the harder
hash).

A "system*last_uid" key was added, managed by exodm_db_system,
in order to increment a UID counter.