Allow JSON-RPC commands to be sent as multipart/form-data
Especially when sending Yang modules, it can be difficult to get
the escaping of quotes right from within curl.

This change lets exodm detect and parse a multipart post, created
e.g. using curl ... -F jsonrpc="..." -F file=@filename

One form part must be named 'jsonrpc', and will be used as the
JSON-RPC request. It is possible to include a file, -F jsonrpc=@...

Files, e.g. the yang spec, can be uploaded using -F file=...
Either enter the content on the command line, or include a file
using @filename. If a different filename is to be specified, use
- F "file=@localfile;filename=nameinpost"

Currently, only the command 'exodm:create-yang-module' recognizes
attached files. If the 'yang-module' attribute is set to
"file:filename", the file "filename" will be located in the list of
files included in the POST.

Example:
$ cat yang.json
{"jsonrpc": "2.0",
 "id": 1,
 "method": "exodm:create-yang-module",
 "params": {"repository":"user",
            "name":"test.yang",
            "yang-module":"file:test.yang"}}

$ curl -u wiger-admin:wiger -k https://localhost:8000/exodm/rpc -F \
   file=@test.yang -F jsonrpc=@yang.json -F file=@yang.json
{"result":{"result":"ok"},"id":1,"jsonrpc":"2.0"}