Implement support for config set areas 'staged', 'pending', 'installed'
When updating a config set, values are stored in the 'staged' area.
When a push-config-set is ordered, the values are cached, and then moved
to 'pending'. For each device, if the push succeeds, it deletes its own
cache entry. The one who is the last one out, switches the 'pending'
area to 'installed'.

The read-config-set-data rpc fetches data from either area, as specified
in the RPC.