User session management and RPC dispatcher threads
Dispatcher threads erroneously checked if devices were  online even
when doing upstream dispatch. Also, they shouldn't rely on the
timeout-protected user sessions, since a session timeout in the
middle of a dispatch can cause unwanted exceptions.

Dispatcher threads now simply identify as the account and user
responsible for the RPC, and are marked as 'trusted processes',
which are allowed to proceed even if the user session isn't active.